requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12
django-cors-headers>=3.7
djangorestframework-simplejwt>=5.0
PyMySQL>=1.0
Pillow>=9.0
PyJWT>=2.0
